# Optimize Categorical.log_prob() of enumerated support