This adds an optimization to `dist.Categorical` to completely avoid tensor ops in `x.log_prob(x.enumerate_support())`, thus speeding up inference using enumeration. In particular it can replace a `torch.gather()` with a `reshape`-`transpose`-`unsqueeze`.

## Performance

- [examples/tabular.py](https://github.com/pyro-ppl/pyro/blob/contrib-treecat/examples/tabular.py): 139s -> 79s
  (that's overall; `torch.gather` was taking 70s alone, all of which is eliminated by this PR)
- examples/hmm.py --model 1: 15.5s -> 13.1s
- examples/hmm.py --model 4: 21.5s -> 18.8s
- examples/hmm.py --model 6: 73.9s -> 32.1s
- examples/hmm.py --model 6 --jit: 63.8s -> 26.4s

## Tested
- refactoring is covered by existing tests, especially test_enum.py
- tested performance on contrib-treecat branch