@fritzo 2019-04-19
> should we consider doing a minor release after merging this?

Sure. I defer to @neerajprad for his minor release supporting MOBB.

> Are there any other changes we are looking to put in?

No changes on my end. @neerajprad may want to put in more changes. The feature I plan to release is a `minipyro.JitTrace_ELBO`, but there is no hurry for that.