@eb8680 2019-04-19
> Are there any other changes we are looking to put in?

Not a release blocker if I don't get around to it, but at some point I would like to update the mixed-effect HMM examples to use `Vindex`, which should dramatically simplify the gross indexing code there.  @fritzo did you want to include an updated version of the enumeration tutorial?  I figure if we're releasing `Vindex` we might as well advertise it a bit.