@fritzo 2019-04-19
> did you want to include an updated version of the enumeration tutorial?

Likewise this is not a release blocker. I'll try to get around to it, but @neerajprad don't wait for me to release.